backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.24.10
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
